fix: change imports when using the splitter_mr library
