+++ .env.template
@@ -6,3 +6,6 @@ AZURE_OPENAI_DEPLOYMENT=XXXX
 
 # OpenAI Keys
 OPENAI_API_KEY=XXXX
+
+# Python path
+PYTHONPATH=.
